Move files in correct directory and fix imports
