+++ lib/controllers/shopware/pageCrawler.js
@@ -1,6 +1,6 @@
 const cheerio = require('cheerio');
 const axios = require('axios');
-const Product = require('../schemas/productSchema');
+const Product = require('../../../schemas/productSchema');
 
 function cleanPrice(str) {
   const removeAfter = str.indexOf('€');
@@ -8,9 +8,10 @@ function cleanPrice(str) {
   return cleanString;
 }
 
-async function scrapeSingleURL(productURL, id) {
-
+async function scrapeURL(productURL, scrapedURLs) {
+  try {
     console.log(productURL);
+    scrapedURLs.push(productURL);
     const res = await axios.get(productURL);
     const $ = cheerio.load(res.data);
     if ($('.product--details').length > 0) {
@@ -25,16 +26,26 @@ async function scrapeSingleURL(productURL, id) {
       const description = $('div.product--description').text();
 
       console.log(title, description, img, price, numPrice);
-      const updatedProduct = {
+      const crawledProduct = new Product({
         url: productURL,
         title: title,
         description: description,
         img: img,
         price: numPrice
-      };
-      console.log(updatedProduct);
-      await Product.updateOne({ _id: id }, updatedProduct);
+      });
+      await crawledProduct.save();
     }
-  
+    const allLinks = $('a');
+    for (let i = 0; i < allLinks.length; i++) {
+      const link = allLinks.get(i);
+      const href = link.attribs.href || '';
+
+      if (href.startsWith('https://shop.sompex.de/') && !scrapedURLs.includes(href)) {
+        await scrapeURL(href, scrapedURLs);
+      }
+    }
+  } catch (error) {
+    console.error('Error: ', productURL, error);
+  }
 }
-exports.scrapeSingleURL = scrapeSingleURL;
+exports.scrapeURL = scrapeURL;
